+++ package.json
@@ -598,10 +598,10 @@
     "@kbn/reporting-export-types-csv": "link:packages/kbn-reporting/export_types/csv",
     "@kbn/reporting-export-types-csv-common": "link:packages/kbn-reporting/export_types/csv_common",
     "@kbn/reporting-export-types-deprecated": "link:packages/kbn-reporting/export_types/deprecated",
-    "@kbn/reporting-export-types-pdf": "link:packages/kbn-reporting/export_types/printable_pdf_v2",
+    "@kbn/reporting-export-types-pdf": "link:packages/kbn-reporting/export_types/pdf",
+    "@kbn/reporting-export-types-pdf-common": "link:packages/kbn-reporting/export_types/pdf_common",
     "@kbn/reporting-export-types-png": "link:packages/kbn-reporting/export_types/png",
     "@kbn/reporting-export-types-png-common": "link:packages/kbn-reporting/export_types/png_common",
-    "@kbn/reporting-export-types-printable_pdf_v2-common": "link:packages/kbn-reporting/export_types/printable_pdf_v2_common",
     "@kbn/reporting-plugin": "link:x-pack/plugins/reporting",
     "@kbn/reporting-server": "link:packages/kbn-reporting/server",
     "@kbn/resizable-layout": "link:packages/kbn-resizable-layout",
