+++ packages/kbn-reporting/export_types/csv_common/index.ts
@@ -6,6 +6,37 @@
  * Side Public License, v 1.
  */
 
-export function foo() {
-  return 'hello world';
+import type { SerializedSearchSourceFields } from '@kbn/data-plugin/public';
+import type {
+  BaseParams,
+  BaseParamsV2,
+  BasePayload,
+  BasePayloadV2,
+} from '@kbn/reporting-common/types';
+
+export interface JobParamsDownloadCSV {
+  browserTimezone: string;
+  title: string;
+  searchSource: SerializedSearchSourceFields;
+  columns?: string[];
 }
+
+interface BaseParamsCSV {
+  searchSource: SerializedSearchSourceFields;
+  columns?: string[];
+}
+
+export type JobParamsCSV = BaseParamsCSV & BaseParams;
+export type TaskPayloadCSV = BaseParamsCSV & BasePayload;
+
+interface CsvFromSavedObjectBase {
+  objectType: 'search';
+}
+
+/**
+ * Makes title optional, as it can be derived from the saved search object
+ */
+export type JobParamsCsvFromSavedObject = CsvFromSavedObjectBase &
+  Omit<BaseParamsV2, 'title'> & { title?: string };
+
+export type TaskPayloadCsvFromSavedObject = CsvFromSavedObjectBase & BasePayloadV2;
