+++ packages/kbn-reporting/export_types/csv_common/tsconfig.json
@@ -13,5 +13,7 @@
   "exclude": [
     "target/**/*"
   ],
-  "kbn_references": []
+  "kbn_references": [
+    "@kbn/data-plugin",
+  ]
 }
