+++ packages/kbn-reporting/export_types/deprecated/index.ts
@@ -7,4 +7,3 @@
  */
 
 export { PdfV1ExportType } from './printable_pdf';
-export type { JobParamsPDFDeprecated } from './printable_pdf';
