+++ packages/kbn-reporting/export_types/pdf/kibana.jsonc
