+++ packages/kbn-reporting/export_types/pdf/printable_pdf_v2.ts
