+++ packages/kbn-reporting/export_types/pdf/tsconfig.json
