+++ packages/kbn-reporting/export_types/pdf_common/README.md
@@ -1,0 +1,3 @@
+# @kbn/reporting-export-types-pdf-common
+
+Empty package generated by @kbn/generate
