+++ packages/kbn-reporting/export_types/pdf_common/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@kbn/reporting-export-types-printable_pdf_v2-common",
+  "name": "@kbn/reporting-export-types-pdf-common",
   "private": true,
   "version": "1.0.0",
   "license": "SSPL-1.0 OR Elastic License 2.0"
