--- packages/kbn-reporting/export_types/printable_pdf_v2_common/README.md
@@ -1,3 +1,0 @@
-# @kbn/reporting-export-types-printable_pdf_v2-common
-
-Empty package generated by @kbn/generate
