+++ packages/kbn-reporting/server/get_full_urls.test.ts
@@ -7,8 +7,8 @@
  */
 
 import { ReportingConfigType, ReportingServerInfo } from '@kbn/reporting-common/types';
+import { TaskPayloadPDF } from '@kbn/reporting-export-types-pdf-common';
 
-import { TaskPayloadPDF } from './types';
 import { getFullUrls } from './get_full_urls';
 
 const getMockJob = (base: object) => base as TaskPayloadPDF;
