+++ tsconfig.base.json
@@ -1180,14 +1180,14 @@
       "@kbn/reporting-export-types-csv-common/*": ["packages/kbn-reporting/export_types/csv_common/*"],
       "@kbn/reporting-export-types-deprecated": ["packages/kbn-reporting/export_types/deprecated"],
       "@kbn/reporting-export-types-deprecated/*": ["packages/kbn-reporting/export_types/deprecated/*"],
-      "@kbn/reporting-export-types-pdf": ["packages/kbn-reporting/export_types/printable_pdf_v2"],
-      "@kbn/reporting-export-types-pdf/*": ["packages/kbn-reporting/export_types/printable_pdf_v2/*"],
+      "@kbn/reporting-export-types-pdf": ["packages/kbn-reporting/export_types/pdf"],
+      "@kbn/reporting-export-types-pdf/*": ["packages/kbn-reporting/export_types/pdf/*"],
+      "@kbn/reporting-export-types-pdf-common": ["packages/kbn-reporting/export_types/pdf_common"],
+      "@kbn/reporting-export-types-pdf-common/*": ["packages/kbn-reporting/export_types/pdf_common/*"],
       "@kbn/reporting-export-types-png": ["packages/kbn-reporting/export_types/png"],
       "@kbn/reporting-export-types-png/*": ["packages/kbn-reporting/export_types/png/*"],
       "@kbn/reporting-export-types-png-common": ["packages/kbn-reporting/export_types/png_common"],
       "@kbn/reporting-export-types-png-common/*": ["packages/kbn-reporting/export_types/png_common/*"],
-      "@kbn/reporting-export-types-printable_pdf_v2-common": ["packages/kbn-reporting/export_types/printable_pdf_v2_common"],
-      "@kbn/reporting-export-types-printable_pdf_v2-common/*": ["packages/kbn-reporting/export_types/printable_pdf_v2_common/*"],
       "@kbn/reporting-plugin": ["x-pack/plugins/reporting"],
       "@kbn/reporting-plugin/*": ["x-pack/plugins/reporting/*"],
       "@kbn/reporting-server": ["packages/kbn-reporting/server"],
