+++ x-pack/plugins/canvas/public/components/workpad_header/share_menu/utils.ts
@@ -6,6 +6,7 @@
  */
 
 import type { RedirectOptions } from '@kbn/share-plugin/public';
+import { JobAppParamsPDFV2 } from '@kbn/reporting-export-types-pdf-common';
 import { CanvasAppLocatorParams, CANVAS_APP_LOCATOR } from '../../../../common/locator';
 import { CanvasWorkpad } from '../../../../types';
