+++ x-pack/plugins/reporting/server/routes/internal/generate/csv_searchsource_immediate.ts
@@ -8,7 +8,8 @@
 import Boom from '@hapi/boom';
 import { schema } from '@kbn/config-schema';
 import type { KibanaRequest, Logger } from '@kbn/core/server';
-import { JobParamsDownloadCSV, ReportingRequestHandlerContext } from '@kbn/reporting-server';
+import { ReportingRequestHandlerContext } from '@kbn/reporting-server';
+import { JobParamsDownloadCSV } from '@kbn/reporting-export-types-csv-common';
 import moment from 'moment';
 import type { ReportingCore } from '../../..';
 import { CSV_SEARCHSOURCE_IMMEDIATE_TYPE, INTERNAL_ROUTES } from '../../../../common/constants';
