+++ x-pack/test/tsconfig.json
@@ -145,7 +145,6 @@
     "@kbn/ecs",
     "@kbn/coloring",
     "@kbn/es",
-    "@kbn/reporting-export-types-png",
     "@kbn/reporting-export-types-deprecated",
     "@kbn/reporting-export-types-csv",
     "@kbn/reporting-server",
